+++ PostGame.h
@@ -5,7 +5,7 @@
 class PostGame : public GameState
 {
 public:
-	explicit PostGame(int score) noexcept;
+	explicit PostGame(size_t score) noexcept;
 
 	void Render(const SpaceInvadersResourceManager& resources) const override;
 	void HandleInput() override;
@@ -19,10 +19,10 @@ private:
 class TransitionToPostGame : public StateChangeTransition
 {
 public:
-	explicit TransitionToPostGame(int score) noexcept;
+	explicit TransitionToPostGame(size_t score) noexcept;
 
 	std::unique_ptr<GameState> ConstructState() const override;
 private:
-	int score;
+	size_t score;
 
 };
